refactor(#138): move _suggest_fix into linting.py; _QUOTED_RE to _core
Finish the lint module: the #29 fix-snippet generator _suggest_fix (and its
_DIAM_MATCH_TOL) moves from make_drawing.py into linting.py alongside
lint_feature_coverage. The shared _QUOTED_RE (a single-quoted-label regex used by
both _suggest_fix and the #30 repair loop) relocates to _core, so linting.py and
make_drawing's repair both import it from below — no cycle.

Pure move: Drawing.lint() calls _suggest_fix via a re-import; tests that import it
from make_drawing keep working. Golden gate passes against the committed snapshots
with no regeneration; suggestion/lint tests pass.
