+++ src/draftwright/linting.py
@@ -1,22 +1,24 @@
-"""Lint-side build state (#138 / ADR 0005, Step 3).
+"""The lint module (#138 / ADR 0005).
 
-ADR 0005 §2 assigns the drawing's *coverage signal* a single owner on the lint
-side: which features a placed pattern callout already documents, and which
-diameters the per-view callout cap dropped. `lint_feature_coverage` consumes this
-to avoid double-reporting a hole that a grouped ``n× ⌀`` callout covers (#92) and
-to suppress the redundant ``feature_not_dimensioned`` for capped diameters.
+Holds the lint-side logic and state:
 
-`Drawing` delegates here and keeps `_pattern_callouts` / `_patterned_holes` /
-`_dropped_callout_diams` reachable as properties during the migration (ADR
-0005 §4). This module is the designated home for the lint functions
-(`lint_feature_coverage`, `_suggest_fix`, scoring) that move out of
-`make_drawing.py` in a later step; for now it owns just the state they read.
+- `lint_feature_coverage` — the completeness check that reports part diameters
+  with no callout (#80), avoiding double-reporting a hole a grouped ``n× ⌀``
+  callout covers (#92) and suppressing the redundant ``feature_not_dimensioned``
+  for capped diameters.
+- `_suggest_fix` — the per-issue ready-to-paste fix snippet (#29).
+- `CoverageState` — the coverage signal the passes record and the checks read
+  (pattern callouts, patterned holes, dropped callout diameters). `Drawing`
+  delegates to it and keeps `_pattern_callouts` / `_patterned_holes` /
+  `_dropped_callout_diams` reachable as properties during the migration (§4).
 
 It sits low in the import DAG — it depends only on `_core` (and build123d_drafting),
-never on `make_drawing`/`annotate`.
+never on `make_drawing`/`annotate`. `Drawing.lint()` calls these via re-imports.
 """
 
 from __future__ import annotations
+
+import re
 
 from build123d_drafting.features import (
     analyse_cylinders,
@@ -25,7 +27,7 @@ from build123d_drafting.features import (
 )
 from build123d_drafting.helpers import LintIssue, TitleBlock
 
-from draftwright._core import _DIAM_RE, _fmt
+from draftwright._core import _DIAM_RE, _QUOTED_RE, _fmt
 
 
 class CoverageState:
@@ -175,3 +177,92 @@ def lint_feature_coverage(
                 )
             )
     return issues
+
+
+# Tolerance for matching a lint message's reported diameter (dedup representative
+# at tol 0.15, formatted to 1 dp) back to a raw feature diameter when generating a
+# fix snippet (#29).
+_DIAM_MATCH_TOL = 0.2
+
+
+def _suggest_fix(issue, dwg) -> str | None:
+    """Return a ready-to-paste code snippet that addresses *issue*, or None.
+
+    The snippet is a hint, not necessarily runnable verbatim (``...`` stands in
+    for args the engine cannot infer). It uses the public domain API
+    (:meth:`Drawing.features`, :meth:`Drawing.at`, :meth:`Drawing.place_dim`)
+    so a caller or LLM can paste and fill the gaps trivially (#29).
+    """
+    code = issue.code
+
+    if code == "feature_not_dimensioned":
+        # Message: "cylindrical feature ø8 has no diameter callout on the sheet".
+        m = _DIAM_RE.search(issue.message)
+        if m is None:
+            return None
+        d = float(m.group(1))
+        # The reported diameter is the dedup representative (tol 0.15) formatted
+        # to 1 dp, so match raw feature diameters with that combined slack — a
+        # 1e-6 match would silently miss every non-integer bore.
+        for view in ("plan", "front", "side"):
+            if any(abs(f.diameter - d) < _DIAM_MATCH_TOL for f in dwg.features(view)):
+                tag = _fmt(d).replace(".", "_")
+                return (
+                    f"# ø{_fmt(d)} has no callout. Locate it via features() and add a leader:\n"
+                    f'for f in dwg.features("{view}"):\n'
+                    f"    if abs(f.diameter - {_fmt(d)}) < {_DIAM_MATCH_TOL}:\n"
+                    f"        callout = HoleCallout(f.diameter, count=f.count,\n"
+                    f"                              through=f.through, depth=f.depth, draft=dwg.draft)\n"
+                    f"        elbow = (f.page_pos[0] + 15, f.page_pos[1] + 10, 0)\n"
+                    f'        leader = Leader((*f.page_pos, 0), elbow, "", dwg.draft, callout=callout)\n'
+                    f'        dwg.add(leader, name="hole_{tag}")'
+                )
+        return None
+
+    if code == "feature_count_mismatch":
+        # Message: "4 ø8 features on the part but callouts account for 1".
+        # `need` is the leading count; anchor it so diameter digits never
+        # interfere regardless of message word order.
+        m = _DIAM_RE.search(issue.message)
+        need_m = re.match(r"\s*(\d+)", issue.message)
+        if m is None or need_m is None:
+            return None
+        need = need_m.group(1)
+        return (
+            f"# Only some ø{m.group(1)} holes are counted. Set count={need} on the "
+            f"callout so it covers them all:\n"
+            f"# HoleCallout(..., count={need}, draft=dwg.draft)"
+        )
+
+    if code == "annotation_overlap":
+        # Message: "labels 'A' and 'B' overlap by ...".
+        labels = _QUOTED_RE.findall(issue.message)
+        first = labels[0] if labels else "<dim>"
+        return (
+            f"# Re-add the dimension with place_dim so it auto-stacks in the "
+            f"layout strip instead of overlapping:\n"
+            f'dwg.remove("{first}")  # if it was named\n'
+            f'dwg.place_dim(p1, p2, "below", "plan", dwg.draft, name="{first}")'
+        )
+
+    if code == "dim_inside_part":
+        # Message: "Dim 'X': annotation bbox overlaps part outline by ...".
+        labels = _QUOTED_RE.findall(issue.message)
+        first = labels[0] if labels else "<dim>"
+        return (
+            f"# The dim sits inside the view — its offset is on the wrong side. "
+            f"Re-place it on the opposite side via place_dim (auto-stacks clear "
+            f"of the part):\n"
+            f'dwg.remove("{first}")  # if it was named\n'
+            f'dwg.place_dim(p1, p2, "right", "front", dwg.draft, name="{first}")'
+        )
+
+    if code == "step_dim_dropped":
+        # Steps too closely spaced to dimension at sheet scale (#41/#42).
+        return (
+            "# Re-build with an enlarged detail view so the crowded shoulders are "
+            "dimensionable:\n"
+            "dwg = build_drawing(part, detail_view=True)"
+        )
+
+    return None
