+++ src/draftwright/make_drawing.py
@@ -82,9 +82,9 @@ from OCP.IFSelect import IFSelect_ReturnStatus
 from OCP.STEPControl import STEPControl_Reader
 
 from draftwright._core import (
-    _DIAM_RE,
     _FONT_SIZE,
     _MARGIN,
+    _QUOTED_RE,
     _SLOT_DIM_HEIGHT,
     _SLOT_DIM_STEP,
     _SLOT_DIM_WIDTH,
@@ -122,7 +122,7 @@ from draftwright.layout import (
     _solve_strip_1d,
     fit_box,
 )
-from draftwright.linting import CoverageState, lint_feature_coverage
+from draftwright.linting import CoverageState, _suggest_fix, lint_feature_coverage
 from draftwright.registry import AnnotationRegistry
 
 _TB_W = 150.0
@@ -475,102 +475,10 @@ _GEOMETRY_AWARE_CODES = frozenset(
 _SCORE_ERROR_PENALTY = 0.2
 _SCORE_WARNING_PENALTY = 0.05
 
-# Quote a label parsed from a lint message safely back into a snippet.
-_QUOTED_RE = re.compile(r"'([^']*)'")
-
 # Lint codes the #30 repair loop can mechanically resolve, and the side flip
 # used to move a dimension that landed on the wrong side of its witness points.
 _REPAIRABLE_CODES = frozenset({"annotation_overlap", "dim_inside_part"})
 _OPPOSITE_SIDE = {"above": "below", "below": "above", "left": "right", "right": "left"}
-
-
-# Tolerance for matching a lint message's reported diameter (dedup
-# representative at tol 0.15, formatted to 1 dp) back to a raw feature
-# diameter when generating a fix snippet (#29).
-_DIAM_MATCH_TOL = 0.2
-
-
-def _suggest_fix(issue, dwg) -> str | None:
-    """Return a ready-to-paste code snippet that addresses *issue*, or None.
-
-    The snippet is a hint, not necessarily runnable verbatim (``...`` stands in
-    for args the engine cannot infer). It uses the public domain API
-    (:meth:`Drawing.features`, :meth:`Drawing.at`, :meth:`Drawing.place_dim`)
-    so a caller or LLM can paste and fill the gaps trivially (#29).
-    """
-    code = issue.code
-
-    if code == "feature_not_dimensioned":
-        # Message: "cylindrical feature ø8 has no diameter callout on the sheet".
-        m = _DIAM_RE.search(issue.message)
-        if m is None:
-            return None
-        d = float(m.group(1))
-        # The reported diameter is the dedup representative (tol 0.15) formatted
-        # to 1 dp, so match raw feature diameters with that combined slack — a
-        # 1e-6 match would silently miss every non-integer bore.
-        for view in ("plan", "front", "side"):
-            if any(abs(f.diameter - d) < _DIAM_MATCH_TOL for f in dwg.features(view)):
-                tag = _fmt(d).replace(".", "_")
-                return (
-                    f"# ø{_fmt(d)} has no callout. Locate it via features() and add a leader:\n"
-                    f'for f in dwg.features("{view}"):\n'
-                    f"    if abs(f.diameter - {_fmt(d)}) < {_DIAM_MATCH_TOL}:\n"
-                    f"        callout = HoleCallout(f.diameter, count=f.count,\n"
-                    f"                              through=f.through, depth=f.depth, draft=dwg.draft)\n"
-                    f"        elbow = (f.page_pos[0] + 15, f.page_pos[1] + 10, 0)\n"
-                    f'        leader = Leader((*f.page_pos, 0), elbow, "", dwg.draft, callout=callout)\n'
-                    f'        dwg.add(leader, name="hole_{tag}")'
-                )
-        return None
-
-    if code == "feature_count_mismatch":
-        # Message: "4 ø8 features on the part but callouts account for 1".
-        # `need` is the leading count; anchor it so diameter digits never
-        # interfere regardless of message word order.
-        m = _DIAM_RE.search(issue.message)
-        need_m = re.match(r"\s*(\d+)", issue.message)
-        if m is None or need_m is None:
-            return None
-        need = need_m.group(1)
-        return (
-            f"# Only some ø{m.group(1)} holes are counted. Set count={need} on the "
-            f"callout so it covers them all:\n"
-            f"# HoleCallout(..., count={need}, draft=dwg.draft)"
-        )
-
-    if code == "annotation_overlap":
-        # Message: "labels 'A' and 'B' overlap by ...".
-        labels = _QUOTED_RE.findall(issue.message)
-        first = labels[0] if labels else "<dim>"
-        return (
-            f"# Re-add the dimension with place_dim so it auto-stacks in the "
-            f"layout strip instead of overlapping:\n"
-            f'dwg.remove("{first}")  # if it was named\n'
-            f'dwg.place_dim(p1, p2, "below", "plan", dwg.draft, name="{first}")'
-        )
-
-    if code == "dim_inside_part":
-        # Message: "Dim 'X': annotation bbox overlaps part outline by ...".
-        labels = _QUOTED_RE.findall(issue.message)
-        first = labels[0] if labels else "<dim>"
-        return (
-            f"# The dim sits inside the view — its offset is on the wrong side. "
-            f"Re-place it on the opposite side via place_dim (auto-stacks clear "
-            f"of the part):\n"
-            f'dwg.remove("{first}")  # if it was named\n'
-            f'dwg.place_dim(p1, p2, "right", "front", dwg.draft, name="{first}")'
-        )
-
-    if code == "step_dim_dropped":
-        # Steps too closely spaced to dimension at sheet scale (#41/#42).
-        return (
-            "# Re-build with an enlarged detail view so the crowded shoulders are "
-            "dimensionable:\n"
-            "dwg = build_drawing(part, detail_view=True)"
-        )
-
-    return None
 
 
 def analyse_face_levels(part, tol: float = 0.5, min_area_frac: float = 0.0) -> list:
